+++ project-base/src/Shopsys/ShopBundle/Model/Order/OrderData.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Order;
+
+use Shopsys\FrameworkBundle\Model\Order\OrderData as BaseOrderData;
+
+class OrderData extends BaseOrderData
+{
+    public function __construct()
+    {
+        parent::__construct();
+    }
+}