+++ project-base/tests/ShopBundle/Database/EntityExtension/Model/ExtendedOrderProduct.php
@@ -33,7 +33,7 @@ class ExtendedOrderProduct extends ExtendedOrderItem
     protected $productStringField;
 
     /**
-     * @param \Shopsys\FrameworkBundle\Model\Order\Order $order
+     * @param \Shopsys\ShopBundle\Model\Order\Order $order
      * @param string $name
      * @param \Shopsys\FrameworkBundle\Model\Pricing\Price $price
      * @param string $vatPercent