+++ project-base/tests/ShopBundle/Database/EntityExtension/Model/ExtendedOrderTransport.php
@@ -5,8 +5,8 @@ namespace Tests\ShopBundle\Database\EntityExtension\Model;
 use Doctrine\ORM\Mapping as ORM;
 use Shopsys\FrameworkBundle\Model\Order\Item\OrderItemData;
 use Shopsys\FrameworkBundle\Model\Order\Item\OrderTransportData;
-use Shopsys\FrameworkBundle\Model\Order\Order;
 use Shopsys\FrameworkBundle\Model\Pricing\Price;
+use Shopsys\ShopBundle\Model\Order\Order;
 use Shopsys\ShopBundle\Model\Transport\Transport;
 
 /**
@@ -35,7 +35,7 @@ class ExtendedOrderTransport extends ExtendedOrderItem
     protected $transportStringField;
 
     /**
-     * @param \Shopsys\FrameworkBundle\Model\Order\Order $order
+     * @param \Shopsys\ShopBundle\Model\Order\Order $order
      * @param string $name
      * @param \Shopsys\FrameworkBundle\Model\Pricing\Price $price
      * @param string $vatPercent